+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to uniplot will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Examples folder.
 
 ## [0.15.1] - 2024-11-03
 ### Fixed